update code before public

+++ Models/Order.cs
@@ -15,7 +15,7 @@ namespace OnlineGasBooking.Models
         public Nullable<int> PaymentID { get; set; }
         public Nullable<int> ShippingID { get; set; }
         public Nullable<int> Discount { get; set; }
-        public Nullable<int> Taxes { get; set; }
+        public Nullable<double> Taxes { get; set; }
         public Nullable<int> TotalAmount { get; set; }
         public Nullable<bool> isCompleted { get; set; }
         public Nullable<System.DateTime> OrderDate { get; set; }
